requirements.txt
cachetools==5.3.1
frozendict==2.3.8
graphviz==0.20.1
isodate==0.6.1
laconia==0.6.0
lxml==4.9.3
PyLD==2.0.3
pyparsing==3.1.1
PyYAML==6.0.1
rdflib==7.0.0
six==1.16.0
SPARQLWrapper==2.0.0
